% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_bin_tree.R
\name{complete.bin.tree}
\alias{complete.bin.tree}
\title{Complete Binary Tree Finder}
\usage{
complete.bin.tree(data.set)
}
\arguments{
\item{dataset}{A dataset of compositional data.  Each row should sum to 1 and
the number of columns should be equal to length of the alpha vector. The
dataset should not contain a header row or identifier column.}
}
\value{
A matrix where every two rows displays the ID of the nodes in the left
branch and right branch of the Dirichlet nesting trees.
}
\description{
Given a compositional dataset, finds the best fitting cascading binary
nested Dirichlet Tree with regards to log likelihood.
}
