% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_like_1.R
\name{log.like.1}
\alias{log.like.1}
\title{Computes the output of the log likelihood function for a Dirichlet
distribution given the alpha parameters and a dataset.}
\usage{
\method{log}{like.1}(alpha, data)
}
\arguments{
\item{alpha}{A vector of alpha parameters for the Dirichlet distribution.}

\item{data}{A dataset of compositional data.  Each row should sum to 1 and
the number of columns should be equal to length of the alpha vector. The
dataset should not contain a header row or identifier column.}
}
\value{
The height of the log likelihood function at the specified alpha
parameters and dataset.
}
\description{
Computes the output of the log likelihood function for a Dirichlet
distribution given the alpha parameters and a dataset.
}
