% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_finder.R
\name{tree.finder}
\alias{tree.finder}
\title{One Step Tree Finder}
\usage{
tree.finder(index, dataset)
}
\arguments{
\item{index}{The index vector gives us the columns to pull from the dataset
indicating which of the parts of the composition to use. A zero indicates
that part of the composition has been dropped.}

\item{dataset}{A dataset of compositional data.  Each row should sum to 1 and
the number of columns should be equal to length of the alpha vector. The
dataset should not contain a header row or identifier column.}
}
\value{
A matrix with two rows displaying the ID of the nodes in the left
branch in the first row and the ID of the nodes in the right branch in the
second row.
}
\description{
Given a compositional dataset, breaks the dataset into two subtrees using the
log-likelihood as the criteria to determine the best bifurcation.  This
function performs one-step in finding the optimal nesting tree.
}
