% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterMaze.R
\docType{data}
\name{WaterMaze}
\alias{WaterMaze}
\title{Proportion of Time Mice Spend in Quadrants of a Water Maze}
\format{
A data frame with 14 observations on 5 variables.
\describe{
  \item{Group}{Whether the mouse was in the healthy group (wild) or a group
  with a mutation that produces symptoms similar to Alzheimer's disease (3tg).}
  \item{TQ}{Proportion of time spent in the target quadrant.}
  \item{AQ1}{Proportion of time spent in adjacent quadrant 1.}
  \item{OQ}{Proportion of time spent in the quadrant opposite the target quadrant. }
  \item{AQ2}{Proportion of time spent in adjacent quadrant 2.}
}
}
\source{
\url{https://github.com/xuod/dirichlet/tree/master/example}
}
\usage{
WaterMaze
}
\description{
Gives the proportion of time each of 14 mice spent in the four quadrants of
a circular Morris water maze.
}
\keyword{datasets}
